log-analyzer-requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
python-dotenv>=1.0.0
google-generativeai>=0.3.1
jinja2>=3.1.2
certifi>=2024.2.2
python-multipart>=0.0.6
openai>=1.0.0
streamlit>=1.27.0
